Stop using exceptions for error handling
As in many scenarios for this daemon exceptions are not used because we
won't be able to handle them, stop using them in the places that we have
added exceptions.
Now instead of crashing the daemon we print an error message and
continue to support other services. Note that a better error handling,
like trying again later is necessary.