Add initial .travis.yml
Using docker with a more recent version of Ubuntu, so we do not need to
maintain a long list of tasks to setup the environment needed to build
Camera Streaming Daemon.

Next step: create a custom image with the dependencies installed, so we
will be able to remove the time-consuming apt-get step.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@intel.com>