Add support to stream cleanup when stream playing has finished
Use calls to media finalize to realize that the media is no longer
playing and that any cleanup needed to be performed by stream class.
Methos finalize_gstreamer_pipeline with the corespondend pipeline will
be called.

This patch also changes the way we do override in gst-rtsp-server
function calls in order to support Glib versions lower than 2.44

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>