stream_builder: Fix an invalid static object access
The current code has a problem in the initialization order of static
objects.
It is not correct to assume that all static objects of a class will be
initialized before another element of that class is accessed. So it is
not safe to access an static object from a class in an object that is
also static.
To fix that, instead of accessing the builders object directly, we're
using a get_builders method that ensures the builders object is created
when it is called.