socket: Change Socket API to include sender/dest addr in read/write
Previous implementation was always using socket write to reply to the
client that made the last request. Instead of this, uses struct
sockaddr_in to identify the client that made the request in read
functions and the destination client in write functions.

This is necessary so we can send packets to multiple clients using a
single socket. For example, using a socket to send broadcast and to
respond client requests.